Extract pure geometry core + behavior-locked unit tests (TUnit net8)
create_walls_from_dwg_poche's pipeline moves to
dwg-commandset/Geometry/PocheGeometryCore.cs: a Revit-free 2D core
(Pt struct) with the handler as a thin adapter. Mazes from the
snapshot: run extraction, jamb sibling classification, face pairing,
poche containment gate, cross-loop collinear merge w/ bridging, and
nested-sliver (pocket slot) filter all live in the core now, owning
their tunables and RejectLog.

New tests/dwg-geometry (net8.0, TUnit) compiles the core by link and
passes 6/6: clean band, room-cavity crosser culled by containment,
sibling-jamb bridging, pocket-door slot strips culled + band
survives, midline-inside-hatch invariant, thickness-band invariant.
Fix a wrap-merge span bug found by the pocket test (span
normalization was missing the +PI the handler used). Add
ElementId((int)) compat for R20-R23 handlers (pre-existing net48
breakage, unrelated to the refactor).

Snapshot before refactor: tag dwg-snapshot-pre-geometry-refactor.

+++ dwg-commandset/Services/Dwg/CreateGridsFromDwgLayerEventHandler.cs
@@ -114,7 +114,7 @@ namespace RevitMCPCommandSet.Services.Dwg
                 double z = DwgCurveSource.MedianZ(allCurves);
                 Level level = null;
                 if (LevelId > 0)
-                    level = doc.GetElement(new ElementId(LevelId)) as Level;
+                    level = doc.GetElement(new ElementId((int)LevelId)) as Level;
                 if (level == null)
                 {
                     level = new FilteredElementCollector(doc)
@@ -257,3 +257,4 @@ namespace RevitMCPCommandSet.Services.Dwg
         }
     }
 }
+
